notifier/requirements.txt
apscheduler==3.10.4
mysql-connector-python==8.4.0
requests==2.31.0
python-dotenv==1.0.1